fix(metriks): force empty Datadog sink hostname. Service name always in prefix.
`Name` maps to Datadog sink `hostname` field. It is only used by
armon/go-metrics lib and in a really non-well documented way.

It is basically used for removing from the metrics key the hostname in
favor of adding it (manually by the user) as tag. The library creator
explains that it makes sense since Datadog is multidimensional so the
tags should be used instead.

See https://github.com/armon/go-metrics/blob/master/datadog/dogstatsd.go#L63-L82

We totally disagree.