Add method to get all enabled feature flags for a user (#187)
In buildbot, we want to start passing feature flags down to netlify build. To do so, we need a method to get all enabled flags for a user. For this particular case, we only care about boolean ones.

This adds a method to the featureflag package to do so. LanuchDarkly's Go SDK provides a way to get all flags for a user, so we wrap that in another method to select just the enabled boolean ones.

For tests, it uses the file config as suggested by LaunchDarkly docs: https://docs.launchdarkly.com/sdk/concepts/flags-from-files