Remove UseDatadog from tracing config
This was introduced in https://github.com/netlify/netlify-commons/pull/247 to allow applications to
use the datadog tracing api directly instead of an opentracing compatible version.

The issue is that when `UseDatadog` is enabled, opentracing.GlobalTracer is set to a noop tracer
so other packages in netlify-commons that emit traces (e.g router) don't work as expected.

We don't actually need this option because `opentracer.New` already calls `tracer.Start` so
applications should be able to use both the opentracing and datadog tracing api.